requirements.txt
Django==2.0.6
django-taggit==0.22.2
pytz==2018.4
